#3
Finished create client router
Added tsconfig.json
Modified db.server.ts
Fixed schame.validation.ts
Added cliented select data to shared folder in API
