+++ src/api/v1/client/client.router.ts
@@ -4,6 +4,7 @@ import { checkSchema, validationResult } from "express-validator"
 import * as ClientService from "./client.service"
 import HttpException from "../../../exception/http.exception";
 import { clientValidation } from "./config/client.validation";
+import { Prisma } from "@prisma/client";
 
 export const clientRouter = express.Router();
 
@@ -15,7 +16,7 @@ clientRouter.get("/:publicId", async (request: Request, response: Response, next
             return response.status(200).json(client)
         next(new HttpException(404, 'client not found'))
     } catch (error: any) {
-        next(new HttpException(404, 'a internal error has occured'))
+        next(new HttpException(500, 'a internal error has occured'))
     }
 })
 
@@ -23,4 +24,13 @@ clientRouter.post("/", checkSchema(clientValidation), async (request: Request, r
     const errors = validationResult(request)
     if (!errors.isEmpty())
         return response.status(400).json({ errors: errors.array() });
+    try {
+        const client = await ClientService.createClient(request.body)
+        return response.status(200).json(client)
+    } catch (error: any) {
+        if (error instanceof Prisma.PrismaClientKnownRequestError)
+            if(error.code === 'P2002')
+                next(new HttpException(400, `The following fields already exists: ${error?.meta?.target.toString()}`))
+        next(new HttpException(500, 'a internal error has occured'))
+    }
 })
